implement comments from review; slow fade-in of events
